test(stripe): split referenceMiddleware describe into middleware.test.ts
Cuts the 9-test referenceMiddleware describe (user + organization
sub-describes) out of stripe.test.ts into its own file so reference
authorization scenarios live next to the corresponding src module.
stripe.test.ts loses ~450 lines.
